+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -21,6 +21,7 @@
 ## Upcoming Release
 * Updated `System.Security.Cryptography.Cng` dependency from `4.4.0` to `5.0.0`.
 * Added Cross region restore support for Delos cloud regions
+* Added Cross region restore support for USSec cloud - USSec West Central region
 
 ## Version 7.12.0
 * Added ChangeSafety Support