+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+
+## Version 7.13.0
 * Updated `System.Security.Cryptography.Cng` dependency from `4.4.0` to `5.0.0`.
 * Added Cross region restore support for Delos cloud regions
 * Updated Az.RecoveryServices.SiteRecovery to use API version 2026-02-01.